backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.9
pydantic-settings>=2.6
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
aiosqlite>=0.20
alembic>=1.14
redis[asyncio]>=5.2
apscheduler>=3.10
litellm>=1.55
anthropic>=0.40
openai>=1.55
httpx>=0.27
tavily-python>=0.5
mcp>=1.0
pyyaml>=6.0
python-ulid>=3.0
python-multipart>=0.0.18
